Update README.md
Recipe Viewer with Favorites is a responsive web application built using Next.js 14+, Tailwind CSS, and MongoDB Atlas, which allows users to discover recipes from TheMealDB API, view detailed ingredients and instructions, and save their favorite meals for easy access. It features a clean UI, favorite recipe management, and is fully optimized for mobile and desktop devices. The project was completed within 24 hours as part of a full-stack development task and is deployed live on Vercel.